turns-service/requirements.txt
flask
requests 
psycopg2-binary